backend/apps/users/views.py
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status

class RegisterView(APIView):
    def post(self, request):
        return Response({"message": "Register endpoint placeholder"}, status=status.HTTP_200_OK)