backend/requirements.txt
Django==5.2.7
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
